Add analysis domain, repos and REST endpoints
Introduce analysis domain types (AnalysisLine, AnalysisResultRecord, AnalysisSession) and several domain exceptions/interfaces. Add repositories for metadata and cached results (AnalysisMetaRepository, AnalysisResultRepository) and update AnalysisGameRepository to persist StartingFen and use the DB Count API (CountByUserIdAsync now returns long). Remove the gRPC Analysis service, add REST DTOs/endpoints and session services, and update tests and feature specs to cover FEN/PGN/match imports, navigation/whatif error handling, and result caching.

+++ Domain/IAnalysisGameRepository.cs
@@ -7,7 +7,7 @@ internal interface IAnalysisGameRepository
     Task<IReadOnlyList<AnalysisGame>> ListByUserIdAsync(
         string userId, int limit, int offset, CancellationToken ct);
 
-    Task<int> CountByUserIdAsync(string userId, CancellationToken ct);
+    Task<long> CountByUserIdAsync(string userId, CancellationToken ct);
 
     Task<AnalysisGame> InsertAsync(AnalysisGame game, CancellationToken ct);
